- Updated object based estimation code using rasterised version of polygons. - Updated commands printed when using "rsgisexe -h algorithm" option. The command rsgislibxmllist can also be used to print commands for all algorithms, with syntax highlighting, and is normally more up to date (as reads from input.xml file).

+++ src/exe/rsgis/RSGISExeTransectModel.h
@@ -94,7 +94,7 @@ protected:
 	options option;
 	unsigned int transectWidth, transectLenght, transectHeight;
 	unsigned int nTrees;
-	unsigned int quadratSize;
+	double quadratSize;
 	unsigned int nRuns;
 	double treeHeight;
 	double transectRes;
